[Refactor] Move table function sink out of Exec
