+++ be/src/exec/data_sinks/data_sink_factory.cpp
@@ -34,9 +34,9 @@
 #include "data_sink/external/mysql_table_sink.h"
 #include "data_sink/external/noop_sink.h"
 #include "data_sink/external/schema_table_sink.h"
+#include "data_sink/external/table_function_table_sink.h"
 #include "data_sink/result/memory_scratch_sink.h"
 #include "data_sink/result/result_sink.h"
-#include "exec/data_sinks/table_function_table_sink.h"
 #include "runtime/runtime_state.h"
 
 namespace starrocks {
