+++ be/src/exec/data_sinks/data_sink_pipeline.cpp
@@ -57,9 +57,9 @@
 #include "data_sink/exchange/multi_cast_data_stream_sink.h"
 #include "data_sink/external/mysql_table_sink.h"
 #include "data_sink/external/noop_sink.h"
+#include "data_sink/external/table_function_table_sink.h"
 #include "data_sink/result/memory_scratch_sink.h"
 #include "data_sink/result/result_sink.h"
-#include "exec/data_sinks/table_function_table_sink.h"
 #include "runtime/runtime_state.h"
 
 namespace starrocks {
